Add clesperantoj prototype wrapper around clesperanto
The problem with wrapping clesperanto directly is that it exposes
Objects which in turn uses vectors, arrays, cl::Memory, and other types that
would also have to be wrapped.  Long term we may do this but it is more
complicated.  So first iteration we create a wrapper class with simpler
interfaces that are more straight forward to wrap.