Add cleperantoj configuration class
This is a 'configuraton class'.  It specifies that we want to wrap
clesperantoj.h and also sets some information in the InfoMap as to how
to wrap the templated functions.

The javacpp 'parser' will use this class to create a java wrapper class
automatically from the header file.  In turn javacpp will invoke the c++
compiler to generate the native wrappers.

See https://github.com/bytedeco/javacpp/wiki/Mapping-Recipes for more
details.