Experimental ObjectJ wrapper approach
Wrapping the CLIc Object class is not trivial as it may involve wrapping
a signficiant portion of opencl.hpp because Object uses classes from
there such as Memory.

A potential workaround is to wrap Object in a class called ObjectJ.
Make the Object member private, but also make ClesperantoJ a friend
class to ObjectJ.  This way java code can have a reference to ObjectJ
and thus keep a copy of the underlying GPU memory to re-use but we can
avoid wrapping Object (and the underlying opencl.hpp classes it uses).
Since ClesperantoJ is a friend of ObjectJ, functions in ClesperantoJ can
access the underlying native Object without having to expose it to Java.